# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroHub API
  description: 'API operacional de AeroHub: aprovisionamiento de tenants, seguimiento
    de vuelos y distribucion de FIDS en tiempo real (RF-T02, OpenAPI 3.1).'
  contact:
    name: AeroHub Platform Team
  version: 0.1.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /vuelos:
    post:
      tags:
      - vuelos
      summary: Crear Vuelo
      description: Da de alta un vuelo para el tenant del llamador (RF-O02 parcial).
      operationId: crear_vuelo_vuelos_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VueloCrearRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VueloCrearResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - vuelos
      summary: Listar Vuelos Endpoint
      description: 'Listado de vuelos del tenant, mas reciente primero, con filtros

        opcionales de fecha y estado actual (Fase 3 de

        docs/diseno/PLAN_CORRECCION_MODULOS.md, causa raiz D): antes de este

        endpoint la vista nucleo no podia mostrar nada al entrar, solo tras

        un evento de WebSocket.'
      operationId: listar_vuelos_endpoint_vuelos_get
      parameters:
      - name: fecha_operacion
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Fecha Operacion
      - name: codigo_estado
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Codigo Estado
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VueloListadoResponse'
                title: Response Listar Vuelos Endpoint Vuelos Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /vuelos/{vuelo_id}:
    get:
      tags:
      - vuelos
      summary: Obtener Vuelo
      description: 'Consulta un vuelo por id (PN-01: 404 si es de otro tenant).'
      operationId: obtener_vuelo_vuelos__vuelo_id__get
      parameters:
      - name: vuelo_id
        in: path
        required: true
        schema:
          type: integer
          title: Vuelo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VueloConsultaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /vuelos/{vuelo_id}/estados:
    post:
      tags:
      - vuelos
      summary: Cambiar Estado Vuelo
      description: 'Registra un cambio de estado y lo publica en el canal en tiempo

        real /vuelos/ws/estado (RF-O04, RNF-P01).'
      operationId: cambiar_estado_vuelo_vuelos__vuelo_id__estados_post
      parameters:
      - name: vuelo_id
        in: path
        required: true
        schema:
          type: integer
          title: Vuelo Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambioEstadoRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CambioEstadoResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /vuelos/informes/simple:
    get:
      tags:
      - vuelos
      summary: Informe Vuelos Simple Endpoint
      operationId: informe_vuelos_simple_endpoint_vuelos_informes_simple_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: aerolinea_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Aerolinea Id
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /vuelos/informes/compuesto:
    get:
      tags:
      - vuelos
      summary: Informe Vuelos Compuesto Endpoint
      operationId: informe_vuelos_compuesto_endpoint_vuelos_informes_compuesto_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CambioEstadoRequest:
      properties:
        codigo_estado_nuevo:
          type: string
          minLength: 1
          title: Codigo Estado Nuevo
        origen_cambio:
          type: string
          title: Origen Cambio
      type: object
      required:
      - codigo_estado_nuevo
      - origen_cambio
      title: CambioEstadoRequest
    CambioEstadoResponse:
      properties:
        vuelo_estado_id:
          type: string
          title: Vuelo Estado Id
      type: object
      required:
      - vuelo_estado_id
      title: CambioEstadoResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    VueloConsultaResponse:
      properties:
        id:
          type: string
          title: Id
        aerolinea_id:
          type: string
          title: Aerolinea Id
        aeronave_id:
          type: string
          title: Aeronave Id
        numero_vuelo:
          type: string
          title: Numero Vuelo
        tipo_vuelo_id:
          type: string
          title: Tipo Vuelo Id
        fecha_operacion:
          type: string
          format: date
          title: Fecha Operacion
        sentido:
          type: string
          title: Sentido
        aeropuerto_origen_id:
          type: string
          title: Aeropuerto Origen Id
        aeropuerto_destino_id:
          type: string
          title: Aeropuerto Destino Id
        sta_utc:
          type: string
          format: date-time
          title: Sta Utc
        std_utc:
          type: string
          format: date-time
          title: Std Utc
        pax_estimado:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pax Estimado
      type: object
      required:
      - id
      - aerolinea_id
      - aeronave_id
      - numero_vuelo
      - tipo_vuelo_id
      - fecha_operacion
      - sentido
      - aeropuerto_origen_id
      - aeropuerto_destino_id
      - sta_utc
      - std_utc
      - pax_estimado
      title: VueloConsultaResponse
    VueloCrearRequest:
      properties:
        aerolinea_id:
          type: integer
          title: Aerolinea Id
        aeronave_id:
          type: integer
          title: Aeronave Id
        numero_vuelo:
          type: string
          maxLength: 10
          minLength: 1
          title: Numero Vuelo
        tipo_vuelo_id:
          type: integer
          title: Tipo Vuelo Id
        fecha_operacion:
          type: string
          format: date
          title: Fecha Operacion
        sentido:
          type: string
          title: Sentido
        aeropuerto_origen_id:
          type: integer
          title: Aeropuerto Origen Id
        aeropuerto_destino_id:
          type: integer
          title: Aeropuerto Destino Id
        sta_utc:
          type: string
          format: date-time
          title: Sta Utc
        std_utc:
          type: string
          format: date-time
          title: Std Utc
        pax_estimado:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pax Estimado
        tenant_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - aerolinea_id
      - aeronave_id
      - numero_vuelo
      - tipo_vuelo_id
      - fecha_operacion
      - sentido
      - aeropuerto_origen_id
      - aeropuerto_destino_id
      - sta_utc
      - std_utc
      title: VueloCrearRequest
    VueloCrearResponse:
      properties:
        vuelo_id:
          type: string
          title: Vuelo Id
      type: object
      required:
      - vuelo_id
      title: VueloCrearResponse
    VueloListadoResponse:
      properties:
        id:
          type: string
          title: Id
        aerolinea_id:
          type: string
          title: Aerolinea Id
        aeronave_id:
          type: string
          title: Aeronave Id
        numero_vuelo:
          type: string
          title: Numero Vuelo
        tipo_vuelo_id:
          type: string
          title: Tipo Vuelo Id
        fecha_operacion:
          type: string
          format: date
          title: Fecha Operacion
        sentido:
          type: string
          title: Sentido
        aeropuerto_origen_id:
          type: string
          title: Aeropuerto Origen Id
        aeropuerto_destino_id:
          type: string
          title: Aeropuerto Destino Id
        sta_utc:
          type: string
          format: date-time
          title: Sta Utc
        std_utc:
          type: string
          format: date-time
          title: Std Utc
        pax_estimado:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pax Estimado
        codigo_estado:
          anyOf:
          - type: string
          - type: 'null'
          title: Codigo Estado
      type: object
      required:
      - id
      - aerolinea_id
      - aeronave_id
      - numero_vuelo
      - tipo_vuelo_id
      - fecha_operacion
      - sentido
      - aeropuerto_origen_id
      - aeropuerto_destino_id
      - sta_utc
      - std_utc
      - pax_estimado
      - codigo_estado
      title: VueloListadoResponse
...